utils: move function to center_info to avoid circular import
