+++ scraper/export/export_merge.py
@@ -6,13 +6,13 @@ from typing import Iterator
 import pytz
 
 from scraper.error import BlockedByDoctolibError
-from scraper.pattern.center_info import CenterInfo, full_dict_to_center
+from scraper.pattern.center_info import CenterInfo, full_dict_to_center, is_reserved_center
 from utils.vmd_blocklist import get_blocklist_urls, is_in_blocklist
 from utils.vmd_center_sort import sort_center
 from utils.vmd_duplicated import deduplicates_names
 from utils.vmd_logger import get_logger
 from utils.vmd_opendata import copy_omit_keys
-from utils.vmd_utils import departementUtils, is_reserved_center
+from utils.vmd_utils import departementUtils
 
 POOL_SIZE = int(os.getenv("POOL_SIZE", 50))
 PARTIAL_SCRAPE = float(os.getenv("PARTIAL_SCRAPE", 1.0))
